service/requirements.txt
pandas
pathlib
openai
tiktoken
fastapi
pydantic
nltk
asyncpg
uvicorn
requests
python-dotenv
psycopg2-binary
faiss-cpu
databases
databases[sqlite]
asyncpg
sqlalchemy
asgiref
langchain
langchain-openai
chromadb